requirements.txt
PySide6>=6.7
requests>=2.32
reportlab>=4.2
pillow>=10.4
qrcode>=7.4
pyinstaller>=6.0
